Remove unnucessary `.pyi`
