--- robyn/events.pyi
@@ -1,3 +1,0 @@
-class Events:
-    STARTUP = ...
-    SHUTDOWN = ...
