--- robyn/ws.pyi
@@ -1,9 +1,0 @@
-from robyn import Robyn
-
-class WS:
-    """This is the python wrapper for the web socket that will be used here."""
-
-    def __init__(self, robyn_object: Robyn, endpoint: str) -> None:
-        pass
-    def on(self, type: str) -> None:
-        pass
